% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_correlations.R
\name{display}
\alias{display}
\title{Display specified correlation table}
\usage{
display(obj, type)
}
\arguments{
\item{obj}{A "tcorr" object.}

\item{type}{Character string; type of the correlation table to display. Explained in detail in \code{\link{tcorr}}.}
}
\value{
None; prints the correlation table.
}
\description{
Display specified correlation table
}
\examples{
\dontrun{
display(result, "HH")
}
}
\seealso{
\code{\link{tcorr}}, \code{\link{tables}}, \code{\link{thresholds}}, \code{\link{statistics}}
}
