% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_correlations.R
\name{statistics}
\alias{statistics}
\title{Compute summary statistics}
\usage{
statistics(obj, type)
}
\arguments{
\item{obj}{A "tcorr" object.}

\item{type}{Character string; type of the correlation table for which to compute statistics.}
}
\value{
List; summary statistics for the specified correlation table, including minimum, maximum, and average correlation values.
}
\description{
Compute summary statistics
}
\examples{
\dontrun{
statistics(result, "HH")
}
}
\seealso{
\code{\link{tcorr}}, \code{\link{tables}}, \code{\link{thresholds}}, \code{\link{display}}
}
