% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_correlations.R
\name{tables}
\alias{tables}
\title{Retrieve specified correlation table}
\usage{
tables(obj, type)
}
\arguments{
\item{obj}{A "tcorr" object.}

\item{type}{Character string; type of the correlation table to retrieve.}
}
\value{
The correlation table matrix.
}
\description{
Retrieve specified correlation table
}
\examples{
\dontrun{
tables(result, "HH")
}
}
\seealso{
\code{\link{tcorr}}, \code{\link{thresholds}}, \code{\link{statistics}}, \code{\link{display}}
}
