% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_correlations.R
\name{tcorr}
\alias{tcorr}
\title{#' @importFrom stats cor}
\usage{
tcorr(
  x,
  y,
  type = c("all", "HH", "LL", "HL", "LH"),
  support = 0.1,
  grid_size = 12,
  threshold_vectors = NULL
)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector; must have the same length as \code{x}.}

\item{type}{Character string; type of correlation table to calculate.
Can be one of "HH", "LL", "HL", "LH", or "all".}

\item{support}{Numeric; minimum percentage of observations needed to report a correlation value.}

\item{grid_size}{Numeric; the size of the grid for the correlation table.}

\item{threshold_vectors}{List; optional custom vectors of thresholds for x and y.}
}
\value{
A list of class "tcorr" containing the correlation tables and threshold vectors.
}
\description{
#' @importFrom stats cor
}
\examples{
\dontrun{
result <- tcorr(runif(100), runif(100))
}
}
\seealso{
\code{\link{tables}}, \code{\link{thresholds}}, \code{\link{statistics}}, \code{\link{display}}
}
