% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_correlations.R
\name{thresholds}
\alias{thresholds}
\title{Retrieve threshold vectors}
\usage{
thresholds(obj)
}
\arguments{
\item{obj}{A "tcorr" object.}
}
\value{
List; the threshold vectors used in \code{obj}.
}
\description{
Retrieve threshold vectors
}
\examples{
\dontrun{
thresholds(result)
}
}
\seealso{
\code{\link{tcorr}}, \code{\link{tables}}, \code{\link{statistics}}, \code{\link{display}}
}
