% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_correlations.R
\name{thrh_index_assoc}
\alias{thrh_index_assoc}
\title{Retrieve threshold index associations}
\usage{
thrh_index_assoc(obj, type)
}
\arguments{
\item{obj}{A "tcorr" object.}

\item{type}{Character string; type of the correlation table.}
}
\value{
Summary of which row/column corresponds to which threshold.
}
\description{
Retrieve threshold index associations
}
\examples{
\dontrun{
thrh_index_assoc(result, "HH")
}
}
\seealso{
\code{\link{tcorr}}, \code{\link{tables}}, \code{\link{thresholds}}, \code{\link{statistics}}, \code{\link{display}}
}
